Update Windows shell functions.txt
Added:
Function: << dea >>
Function shortcut for deactivate; for python virtual environments.

Modified: 
Function << cvenv >> | now << avenv >> 
This change was made to better represent "activate venv" instead of cd'ing to it.

<< avenv >> now also uses Push-Locaiton, and popd to return as they are native functions and don't require a call.

cd2 function still remains as it's overall useful. Previously << cvenv >> function used cs2 to return to the previous location.

Fixed:
An oversight on the previous commit led to some commands being accidentally changed to Set-Location, instead of Push-Location.
This was not intentional and the popd functionality has been restored.